split the shaderresource
